Few fixes to blog
Looks great! When a web crawler indexes your page, priorities are assigned based on your <h*> tags. So, the higher the h (<h1>), the crawler thinks it is more important (and it should be). So, I flipped around some of the <h> tags to make it more semantically correct. If you want to change how it is presented, change your styles for the <h> tags.